Cleanup work to enable feature gating annotations
