+++ plugin/pkg/scheduler/algorithm/predicates/BUILD
@@ -47,6 +47,7 @@ go_test(
         "//vendor:k8s.io/apimachinery/pkg/api/resource",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
         "//vendor:k8s.io/apimachinery/pkg/labels",
+        "//vendor:k8s.io/apiserver/pkg/util/feature",
     ],
 )
